The timing subsystem has been implemented
All slots that deal with the control of the simulation (e.g. play())
have been implemented. The member that holds the view for the lattice
has been added to the controller class. The members that hold the
cellular automaton and the view are now deleted in the destructor of
the controller class.