Fix the movement of predators at the borders
The movement of particles in regioes close to the borders of the lattice
had a serious bug, such issue provoked that a particle that ideally
would cross the border to avoid a long trip to a new position, travels
along the longest path to its new position. The fix is simple: just
validate the magnitude of each vector calculated in the equation for the
new velocity, and if the magnitude of a component is greater that the
corresponding dimension then adjust it according to a relative position.